requirements.txt
pygame-ce>=2.5.0
gymnasium>=0.29.0
stable-baselines3>=2.3.0
sb3-contrib>=2.3.0
torch>=2.1.0
numpy>=1.24.0
tensorboard>=2.14.0
